+++ app/app.vue
@@ -1,7 +1,17 @@
 <template>
   <div>
-    <NuxtLayout >
+    <title>Robaloxs</title>
+    <NuxtLayout>
       <NuxtPage />
     </NuxtLayout>
   </div>
 </template>
+
+<script setup>
+useSeoMeta({
+  title: 'Robaloks - Beli Robux Murah',
+  ogTitle: 'Robaloks - Beli Robux Murah', 
+  description: 'Top up aman dan ori.',
+  ogDescription: 'Top up aman dan ori.',
+})
+</script>