+++ app/components/checkboxes/Default.vue
@@ -1,7 +1,7 @@
 <template>
   <div :class="[base, isActive ? focus : defaultClass]" @click="select">
     <Icon :name="isActive ? 'mdi:radiobox-marked' : 'mdi:radiobox-blank'" size="12" />
-    <img src="@/assets/images/robux.png" alt="Robux" class="w-6" />
+    <img :src="image" alt="Robux" class="w-6" />
     <div class="flex flex-col items-start justify-center">
       <p class="font-inter text-sm font-bold">
         {{ qty }}
@@ -19,6 +19,7 @@ const props = defineProps<{
   value: string
   qty?: string
   price?: string
+  image?: string
 }>()
 
 const emit = defineEmits<{